Tweaks to columns and padding.
I believe that the `span` and `offset` classes should only be used
on divs within `row` divs - this was appearing off centre to me.

My change has made the site a bit wider, but I think that might
actually be nicer, for the moment at least.

I've also added padding between the sections, and to the bottom, so
that the content doesn't site on the bottom of the browser window.
This also seems to have improved the 'change nav highlighting'
mechanism, which previously waited until the relevant heading was
off the screen before switching.